cmd/tailscale/cli: Updated use of service name.
This commit removes serviceName.IsEmpty and use direct comparison to  instead. In legacy code, when an empty service
name needs to be passed, a new constant noService is passed. Removed redundant code for checking service name validity
and string method for serviceNameFlag. Also changed SetByUser field in bgBoolVar to IsSet

Signed-off-by: KevinLiang10 <37811973+KevinLiang10@users.noreply.github.com>