cmd/k8s-proxy: fix error on watch connection killed
k8s-proxy was reliably crashing after a long but variable period of
time. On debugging, this was due to the watcher getting stopped when the
connection was severed (for example if my laptop was asleep for a
while), and the switch-case failed to account for the fact the result
chan may have been closed. Fix by checking for channel being closed, and
restarting the watcher in that case. Also set a timeout on the watch so
we're not relying on the watcher being quick to react to severed
connections and regularly exercise the logic to re-watch. Finally, pull
the loadConfigFromSecret call inside the added/created case to ensure
it's never possible to call it with a nil secret.

Change-Id: Ibf0cf971cbe41524cf6db3c0644c29803ad69fc4
Signed-off-by: Tom Proctor <tomhjp@users.noreply.github.com>