fix up remaining possible sites of optimistic lock errors
Because there are two reconcilers editing the same resource in quick
succession for kube-apiserver, we end up with a lot of optimistic lock
errors, so ensure we catch them all and just quietly retry rather than
logging error messages and events etc.

In future, we should definitely consider these reconciler loops as a
prime candidate for switching to server side apply patches to reduce the
churn here.

Change-Id: Ie09a7e0515568e0e04ab6cf6bdc5bb2f1bf47502
Signed-off-by: Tom Proctor <tomhjp@users.noreply.github.com>