net/netmon: Use OpenVPN interface on Mac
If we are  on a machine where OpenVPN has brought an interface up we
want to explicitly use it by preference to identifying the system's idea
of the default route.

We identify OpenVPN by the combination of a utunN device being present,
and it having routes 0.0.0.0/1 and 128.0.0.0/1, being the two `def1`
half routes that the OpenVPN client installs on Macs. This allows
Tailscale to reach the internet and then function as an overlay
network.

Closes #16652

Signed-off-by: Andrew Cowie <andrew.cowie@apkudo.com>