cmd/k8s-operator: grant secret permissions to auth proxy
This commit modifies the role provided to the `kube-apiserver-auth-proxy`
service account used by the operator when running the kube proxy to
include the ability to get & list secrets.

The proxy attempts to check for the presence of a TLS certificate stored
in a secret but fails due to lacking this permission. This causes a
kubernetes API error to be surfaced in the machines UI in the admin
panel. While the certificate is not required, it may confuse customers
to see "Invalid Certificate" with a heavy kubernetes error.

Signed-off-by: David Bond <davidsbond93@gmail.com>