ipn/ipnlocal: fix the base case of LockedOnEntry use
There are several methods within the LocalBackend that use an unusual and
error-prone lock discipline whereby they require the caller to hold the backend
mutex on entry, but release it on the way out.

PR 11650 added some support code to make this pattern more visible.
Now it is time to eliminate the pattern (at least within this package).

As a first step, the base case for all of this is the state machine transition,
which unlocks after setup. Replace its explicit unlock with an unlock/lock
closure to push responsibility one level higher in the stack.

Updates (#)11649 [fix later, I don't want to spam the issue while testing]

Change-Id: I532856ab9c392ec349b3050f31c64699aca964dd
Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>