ipn/ipnlocal: revert some locking changes ahead of release branch cut
To reduce risk to the shortly-upcoming release, revert some of the incremental
changes to locking behaviour in the LocalBackend. We can reapply these after
the release to get exercise in unstable builds, but for now the schedule is
close.

This reverts the following commits:

b411ffb52f13 "ipn/ipnlocal: remove UnlockEarly from doSetHostinfoFilterServices"
9002e5fd6b8e "ipn/ipnlocal: remove an unnecessary unlock shortcut"
2fb9472990ea "ipn/ipnlocal: remove unnecessary usees of lockAndGetUnlock"
6c8fef961eab "ipn/ipnlocal: replace the LockedOnEntry pattern with conventional lock/unlock discipline (#16925)"

Updates #11649
Updates #15160

Change-Id: I6b1006c18fd1c2f1abfb847761e480a60957b883
Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>