client/tailscale/apitype: add structures to support new /dns/composite endpoint
The /dns/composite endpoint backed by these new structures is meant to support
all possible DNS configurations in a single place. It is similar to
/dns/config (undocumented) except that the field names and structure
are designed to be user-facing, intending to match the experience of
other endpoints and the UI panel as closely as possible.

More immediately, it gives a place in the public API to expose
useWithExitNode in resolvers. The current finer-grained endpoints
(/dns/nameservers, e.g.) can't support this flag or richer configuration
because resolvers are treated as address strings instead of objects.

DNSResolver.Addr is changed to DNSResolver.Address since now the field
name will be user-facing. This breaks consumers of /dns/config, but
since the only consumer is the admin panel, the break can be accounted
for.

DNSResolver.UseWithExitNode is no longer omitempty as the API should
marshal an explicit value in responses.

Updates tailscale/corp#32129
Updates #8237

Signed-off-by: Michael Ben-Ami <mzb@tailscale.com>