tstest/integration: expand the tests for `tailscale up`
Expand the integration tests to cover a wider range of scenarios, including:

*   Auth URLs and auth keys
*   With and without the `--force-reauth` flag
*   Before and after a successful initial login

These tests expose a race condition when using `--force-reauth` on an
already-logged in device. The command completes too quickly, preventing
the auth URL from being displayed. This issue is identified and will be
fixed in the next commit.

Fixes #17108