appc: check published events in unit tests
Update the unit tests in the appc package to also check the events published
alongside calls to the RouteAdvertiser.

For now the tests still rely on the RouteAdvertiser for correctness; this is OK
for now as the two methods are always performed together.  In the near future,
we need to rework the tests so not require that, but that will require building
some more test fixtures that we can handle separately.

Updates #15160
Updates #17192

Change-Id: Iea3dd88af0437b50204e5a7488c6a74d26cebf81
Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>